Started implementing the second part of the tokeniser
I have started implementing a object tokeniser to convert the tokenised
expressions into collection of object tokens, for instance a string
tokenised into 3 parts (e.g. ", ABCD, ") is converted into a string token
which only holds the contents of the string so that matching a pattern
involving a string is less likely to detect incorrect matches and requires
less work to match against. Also this means that matching can be made more
generalised. For example a function call such as foo("bar", 10) can be
broken up into a word object (the name of the call) followed by a open
brace token and can scan until a closing brace token is found for tokens, which
in this case will find a string token and a number token, however without
object tokenisation the compiler would not be able to detect that a string
was used due to the complexity of extracting the token meaning that only simple types such as variables or numbers
could be used in function calls.